Update component folder
