+++ lib/components/assets/svg/battery_svg.dart
