+++ lib/components/assets/svg/people_svg.dart
