+++ lib/components/assets/svg/word_svg.dart
