+++ lib/components/profile/badges_card.dart
