+++ lib/components/profile/token_card.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
 // Components Imports
-import 'package:langx_flutter/components/trophy_svg.dart';
-import 'package:langx_flutter/components/info_svg.dart';
+import 'package:langx_flutter/components/assets/svg/trophy_svg.dart';
+import 'package:langx_flutter/components/assets/svg/info_svg.dart';
 
 class ProfileTokenCard extends StatelessWidget {
   final String imgUrl;
